tab-box: Support Block API Version 2
