+++ blocks/src/block/tab-box/block.json
@@ -1,0 +1,28 @@
+{
+  "apiVersion": 2,
+  "name": "cocoon-blocks/tab-box-1",
+  "category": "cocoon-block",
+  "keywords": [ "tab", "box" ],
+
+  "attributes": {
+    "label": {
+      "type": "string",
+      "default": "bb-check"
+    },
+    "backgroundColor": {
+      "type": "string"
+    },
+    "textColor": {
+      "type": "string"
+    },
+    "borderColor": {
+      "type": "string"
+    },
+    "customBorderColor": {
+      "type": "string"
+    },
+    "fontSize": {
+      "type": "string"
+    }
+  }
+}
