+++ blocks/src/blocks.js
@@ -17,6 +17,7 @@ import * as iconBox from './block/icon-box';
 import * as iconList from './block/icon-list';
 import * as infoBox from './block/info-box';
 import * as stickyBox from './block/sticky-box';
+import * as tabBox from './block/tab-box';
 
 const cocoonBlocks = [
   blankBox,
@@ -24,6 +25,7 @@ const cocoonBlocks = [
   iconList,
   infoBox,
   stickyBox,
+  tabBox,
 ];
 
 
@@ -70,7 +72,7 @@ registerCocoonBlocks();
 //import './block/info-box/block.js';
 //import './block/blank-box/block.js';
 //import './block/sticky-box/block.js';
-import './block/tab-box/block.js';
+//import './block/tab-box/block.js';
 import './block/balloon/block.js';
 //import './block/balloon-ex/block.js';
 import './block/blogcard/block.js';
